refactor(bayn): simplify Effect runtime composition
